generator/requirements.txt
psycopg2-binary==2.9.7
Faker==20.1.0